backend/requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.25.0
sqlalchemy==2.0.23
pydantic==2.7.0
python-multipart==0.0.9
alembic==1.13.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
bcrypt==5.0.0
python-jose[cryptography]==3.3.0
anyio==4.7.0
pydantic-settings==2.2.1
pandas==2.2.2
numpy==2.4.2
yfinance==0.2.36
ccxt==4.2.25
matplotlib>=3.9.0
scipy==1.17.0
